Front services posts and filter v1

+++ src/pages/api/socket.ts
@@ -5,12 +5,12 @@ import { Server as NetServer } from "http";
 import { prisma } from "@/server/lib/prisma";
 import ChatController from "@/server/controllers/chat.controller";
 import AuthMiddleware from "@/server/middlewares/auth.middleware";
+import AuthHelper from "@/server/helpers/auth.helper";
 
 export default function handler(req: NextApiRequest, res: NextApiResponseServerIO) {
 
   const middlewareResponse = AuthMiddleware.checkAuth(req, ['ADMIN', 'CUSTUMER', 'SUPPLIER']);
   if (!middlewareResponse.success) return res.status(401).json({ msg: 'Unauthorized.' });
-  const { user_id } = middlewareResponse
 
   if (!res.socket.server.io) {
 
@@ -18,14 +18,29 @@ export default function handler(req: NextApiRequest, res: NextApiResponseServerI
     const io: ServerIO = new ServerIO(httpServer)
 
     io.on('connection', async (socket) => {
-      let online = await prisma.online.findFirst({ where: { user_id: Number(user_id) } })
-      if (online) {
-        online = await prisma.online.update({ where: { id: online.id }, data: { socket_id: socket.id } })
-      } else {
-        online = await prisma.online.create({ data: { user_id: Number(user_id), socket_id: socket.id } })
-      }
+      let autenticated = false
+      let user_id: number
+      let online: any
+
+      socket.on('authenticate', async (accessToken) => {
+        const authResponse = AuthHelper.authenticateUserToken(accessToken, ['ADMIN', 'CUSTUMER', 'SUPPLIER'])
+        if (authResponse.success) {
+          autenticated = true
+          user_id = authResponse.data?.user_id
+          online = await prisma.online.findFirst({ where: { user_id } })
+          if (online) {
+            online = await prisma.online.update({ where: { id: online.id }, data: { socket_id: socket.id } })
+          } else {
+            online = await prisma.online.create({ data: { user_id, socket_id: socket.id } })
+          }
+          socket.emit('authenticated', true)
+        } else {
+          socket.emit('authenticated', false)
+        }
+      })
 
       socket.on('message', async (msg, recipientId) => {
+        if (!autenticated) return
         const recipient = await prisma.online.findFirst({ where: { user_id: recipientId } })
         if (recipient) {
           io.to(recipient.socket_id).emit('message', msg, socket.id);
@@ -34,6 +49,7 @@ export default function handler(req: NextApiRequest, res: NextApiResponseServerI
       })
 
       socket.on('disconnect', () => {
+        if (!autenticated) return
         prisma.online.delete({ where: { id: online.id } })
       });
     });
